Data is now inserted into database recursively, using `sqla_mapping` dictionary from `DeviceModel`
