+++ tests/fixtures/ttp/huawei-vrp.json
@@ -1,7 +1,8 @@
 {
-    "interfaces": {
-        "Ethernet1/0/0": {
+    "interfaces": [
+        {
+            "name": "Ethernet1/0/0",
             "description": "FIXTURE_TTP_TEST"
         }
-    }
+    ]
 }
