+++ tests/fixtures/ttp/juniper-junos.json
@@ -1,11 +1,13 @@
 {
-    "interfaces": {
-        "ge-0/0/1": {
+    "interfaces": [
+        {
+            "name": "ge-0/0/1",
             "description": "FIXTURE_TTP_TEST"
         },
-        "ge-0/0/1.101": {
+        {
+            "name": "ge-0/0/1.101",
             "description": "FIXTURE_TTP_TEST",
             "s_vlan": "101"
         }
-    }
+    ]
 }
